+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
 
+## Unreleased
+
+### Fix
+
+* Cleaner shutdown. [Ben Dalling]
+
+* Provide the destination when marking sending as failed. [Ben Dalling]
+
+### Performance
+
+* Only parse JSON messages once. [Ben Dalling]
+
+
 ## 0.9.0 (2025-07-23)
 
 ### Features